refactor: watcher for toggling companion
